requirements.txt
elevenlabs
manim
python-dotenv
asyncio
playwright
requests
beautifulsoup4
